requirements.txt
mido>=1.3.0
python-rtmidi>=1.5.0
pyserial>=3.5
